+++ agent/engine/execcmd/manager_start_windows.go
@@ -1,19 +1,12 @@
 package execcmd
 
-import (
-	"context"
+import apicontainer "github.com/aws/amazon-ecs-agent/agent/api/container"
 
-	apicontainer "github.com/aws/amazon-ecs-agent/agent/api/container"
-	apitask "github.com/aws/amazon-ecs-agent/agent/api/task"
-	"github.com/aws/amazon-ecs-agent/agent/dockerclient/dockerapi"
+const (
+	execAgentCmdUser   = "NT AUTHORITY\\SYSTEM"
+	execAgentCmdBinDir = "C:\\Program Files\\Amazon\\SSM"
 )
 
-// Note: exec cmd agent is a linux/windows feature, thus implemented here as a no-op.
-func (m *manager) RestartAgentIfStopped(ctx context.Context, client dockerapi.DockerClient, task *apitask.Task, container *apicontainer.Container, containerId string) (RestartStatus, error) {
-	return NotRestarted, nil
-}
-
-// Note: exec cmd agent is a linux/windows feature, thus implemented here as a no-op.
-func (m *manager) StartAgent(ctx context.Context, client dockerapi.DockerClient, task *apitask.Task, container *apicontainer.Container, containerId string) error {
-	return nil
+func getexecAgentCmdBinDir(ma *apicontainer.ManagedAgent) string {
+	return execAgentCmdBinDir
 }
